Create module localstack-test
