+++ java/common/common-task/src/test/java/sleeper/task/common/QueueMessageCountIT.java
@@ -20,17 +20,11 @@ import com.amazonaws.services.sqs.AmazonSQS;
 import com.amazonaws.services.sqs.AmazonSQSClientBuilder;
 import com.amazonaws.services.sqs.model.QueueDoesNotExistException;
 import org.junit.jupiter.api.Test;
-import org.testcontainers.DockerClientFactory;
 import org.testcontainers.containers.localstack.LocalStackContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.utility.DockerImageName;
 
-import sleeper.core.CommonTestConstants;
-
-import java.io.UncheckedIOException;
-import java.net.InetAddress;
-import java.net.UnknownHostException;
+import sleeper.localstack.test.SleeperLocalStackContainer;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -41,9 +35,7 @@ class QueueMessageCountIT {
     private static final String TEST_QUEUE_NAME = "test-queue-url";
 
     @Container
-    public static LocalStackContainer localStackContainer = new LocalStackContainer(DockerImageName.parse(CommonTestConstants.LOCALSTACK_DOCKER_IMAGE))
-            .withServices(LocalStackContainer.Service.SQS)
-            .withEnv("LOCALSTACK_HOST", getHostAddress());
+    public static LocalStackContainer localStackContainer = SleeperLocalStackContainer.create(LocalStackContainer.Service.SQS);
 
     private AmazonSQS createSQSClient() {
         return buildAwsV1Client(localStackContainer, LocalStackContainer.Service.SQS, AmazonSQSClientBuilder.standard());
@@ -51,15 +43,6 @@ class QueueMessageCountIT {
 
     private String createQueue(AmazonSQS sqs) {
         return sqs.createQueue(TEST_QUEUE_NAME).getQueueUrl();
-    }
-
-    private static String getHostAddress() {
-        String dockerHost = DockerClientFactory.instance().dockerHostIpAddress();
-        try {
-            return InetAddress.getByName(dockerHost).getHostAddress();
-        } catch (UnknownHostException e) {
-            throw new UncheckedIOException(e);
-        }
     }
 
     @Test
